requirements/dev.txt
djangorestframework==3.14.0
django-redis==5.2.0
psycopg2-binary==2.9.6
model-bakery==1.12.0
pytest==7.3.1
pytest-django==4.5.2